requirements.txt
torch>=2.0.0
torchvision
geffnet
tqdm
pillow
